+++ API/.env
@@ -1,0 +1,7 @@
+OPT_PATH = /opt
+PROGRAM_PATH = ${OPT_PATH}/coias-back-app
+IMAGES_PATH = ${OPT_PATH}/tmp_images
+FILES_PATH = ${OPT_PATH}/tmp_files
+SUBARU_PATH = ${PROGRAM_PATH}/SubaruHSC
+DOC_IMAGE_PATH = ${PROGRAM_PATH}/docs/image
+COIAS_PARAM_PATH = /param
