+++ API/test_main.py
@@ -2,8 +2,19 @@ from main import app
 
 import pathlib
 from fastapi.testclient import TestClient
+import config
 
 client = TestClient(app)
+
+
+def test_get_all_path():
+    print(config.COIAS_PARAM_PATH)
+    print(config.DOC_IMAGE_PATH)
+    print(config.FILES_PATH)
+    print(config.IMAGES_PATH)
+    print(config.OPT_PATH)
+    print(config.PROGRAM_PATH)
+    print(config.SUBARU_PATH)
 
 
 def test_get_unknown_disp():
@@ -24,3 +35,7 @@ A B C D
     response = client.get("/unknown_disp")
     assert response.status_code == 200
     assert response.json() == {"result": [["A", "B", "C", "D"], ["0", "1", "2", "3"]]}
+
+
+if __name__ == "__main__":
+    test_get_all_path()
