+++ README.md
@@ -3,6 +3,20 @@
 ```
 .
 ├── API
+│   ├── __init__.py      # "API"をパッケージにするファイル
+│   ├── main.py          # "main"モジュール, e.g. import API.main
+│   ├── dependencies.py  # "dependencies"モジュール, e.g. import API.dependencies
+│   ├── config.py        # "config"モジュール, e.g. import API.config
+│   ├── utils.py         # "utils"モジュール, e.g. import API.utils
+│   └── routers          # "routers" はサブパッケージ
+│   │   ├── __init__.py  # "routers" をサブパッケージにするファイル
+│   │   ├── files.py     # "files"はサブモジュール, e.g. import API.routers.files
+│   │   ├── processes.py # "processes"はサブモジュール, e.g. import API.routers.processes
+│   │   ├── tests.py     # "tests"はサブモジュール, e.g. import API.routers.tests
+│   │   └── ws.py        # "ws"はサブモジュール, e.g. import API.routers.ws
+│   └── internal         # "internal"はサブパッケージ
+│       ├── __init__.py  # makes "internal" a "Python subpackage"
+│       └── admin.py     # "admin" はサブモジュール, e.g. import API.internal.admin
 ├── Docker
 ├── SubaruHSC
 ├── data
